# projects/mark-ms/lib/swagger.yaml
...
paths:
  /marks:
    get:
      summary: Retrieve Marks
      tags:
        - Marks
      responses:
        '200':
          $ref: '#/responses/200 List of Marks'
          
        '400':
          $ref: '#/responses/400 Bad Request'
          
        '401':
          $ref: '#/responses/401 Unauthorized'
          
        '403':
          $ref: '#/responses/403 Forbidden'
          
        '422':
          $ref: '#/responses/422 Unprocessable Entity'
          
        '429':
          $ref: '#/responses/429 Too Many Requests'
          
        '500':
          $ref: '#/responses/500 Server Error'
          
      parameters:
        - $ref: '#/parameters/IDS'
          
        - $ref: '#/parameters/Query Limit'
          
        - $ref: '#/parameters/Continuation Token'
          
    put:
      summary: Revise a Mark
      tags:
        - Marks
      responses:
        '200':
          description: Post was made successfully
          x-oad-type: response
        '400':
          $ref: '#/responses/400 Bad Request'
        '401':
          $ref: '#/responses/401 Unauthorized'
        '403':
          $ref: '#/responses/403 Forbidden'
        '422':
          $ref: '#/responses/422 Unprocessable Entity'
        '429':
          $ref: '#/responses/429 Too Many Requests'
        '500':
          $ref: '#/responses/500 Server Error'
      parameters:
        - $ref: '#/parameters/Revision Number'
        - $ref: '#/parameters/Content'
        - $ref: '#/parameters/Timestamp'
        - $ref: '#/parameters/Key'
    post:
      summary: Post a new Mark
      tags:
        - Marks
      responses:
        '200':
          description: Post was made successfully
          x-oad-type: response
        '400':
          $ref: '#/responses/400 Bad Request'
          
        '401':
          $ref: '#/responses/401 Unauthorized'
          
        '403':
          $ref: '#/responses/403 Forbidden'
          
        '422':
          $ref: '#/responses/422 Unprocessable Entity'
          
        '429':
          $ref: '#/responses/429 Too Many Requests'
          
        '500':
          $ref: '#/responses/500 Server Error'
          
      parameters:
        - $ref: '#/parameters/Key'
          
        - $ref: '#/parameters/Content'
          
        - $ref: '#/parameters/Timestamp'
          
    parameters:
      - name: Content-Type
        in: header
        description: Response type
        required: true
        type: string
        enum:
          - application/json
        x-oad-type: parameter
      - name: Authorization
        in: header
        description: User access token
        required: true
        type: string
        x-oad-type: parameter
    x-oad-type: operation
  '/marks/:tag':
    get:
      summary: Retrieve Marks associated with a tag
      description: |-
        Retrieve marks assocaite with a tag. 
        Marks can come back in orders specifed. Default order is cronological.
      tags:
        - Marks
        - Tags
      responses:
        '200':
          $ref: '#/responses/200 List of Marks'
          
        '400':
          $ref: '#/responses/400 Bad Request'
          
        '401':
          $ref: '#/responses/401 Unauthorized'
          
        '403':
          $ref: '#/responses/403 Forbidden'
          
        '422':
          $ref: '#/responses/422 Unprocessable Entity'
          
        '429':
          $ref: '#/responses/429 Too Many Requests'
          
        '500':
          $ref: '#/responses/500 Server Error'
          
      parameters:
        - $ref: '#/parameters/Tag'
          
        - $ref: '#/parameters/Query Limit'
          
        - $ref: '#/parameters/Continuation Token'
          
    parameters:
      - name: Content-Type
        in: header
        description: Response type
        required: true
        type: string
        enum:
          - application/json
        x-oad-type: parameter
      - name: Authorization
        in: header
        description: User access token
        required: true
        type: string
        x-oad-type: parameter
    x-oad-type: operation
  '/users/:handle/marks':
    get:
      summary: Retrieve Marks by user handle.
      tags:
        - Marks
        - User Accounts
      responses:
        '200':
          $ref: '#/responses/200 List of Marks'
          
        '400':
          $ref: '#/responses/400 Bad Request'
          
        '401':
          $ref: '#/responses/401 Unauthorized'
          
        '403':
          $ref: '#/responses/403 Forbidden'
          
        '422':
          $ref: '#/responses/422 Unprocessable Entity'
          
        '429':
          $ref: '#/responses/429 Too Many Requests'
          
        '500':
          $ref: '#/responses/500 Server Error'
          
      parameters:
        - $ref: '#/parameters/Handle Path'
          
        - $ref: '#/parameters/Query Limit'
          
        - $ref: '#/parameters/Continuation Token'
          
    parameters:
      - name: Content-Type
        in: header
        description: Response type
        required: true
        type: string
        enum:
          - application/json
        x-oad-type: parameter
      - name: Authorization
        in: header
        description: User access token
        required: true
        type: string
        x-oad-type: parameter
    x-oad-type: operation
  '/users/:handle/marks/:id':
    get:
      summary: Retrieve a single Mark by a specified user.
      tags:
        - Marks
      responses:
        '200':
          $ref: '#/responses/200 Single Mark'
          
        '400':
          $ref: '#/responses/400 Bad Request'
          
        '401':
          $ref: '#/responses/401 Unauthorized'
          
        '403':
          $ref: '#/responses/403 Forbidden'
          
        '422':
          $ref: '#/responses/422 Unprocessable Entity'
          
        '429':
          $ref: '#/responses/429 Too Many Requests'
          
        '500':
          $ref: '#/responses/500 Server Error'
          
      parameters:
        - $ref: '#/parameters/Handle Path'
          
        - $ref: '#/parameters/ID'
          
    parameters:
      - name: Content-Type
        in: header
        description: Response type
        required: true
        type: string
        enum:
          - application/json
        x-oad-type: parameter
      - name: Authorization
        in: header
        description: User access token
        required: true
        type: string
        x-oad-type: parameter
    x-oad-type: operation
...
parameters:
...
  ID:
    name: id
    in: path
    description: Mark ID.
    required: true
    type: string
    x-oad-type: parameter
  IDS:
    name: ids
    in: query
    description: >-
      Request for specific IDs.

      Can be left out to query for all most recent without specifying specific
      IDs.
    required: false
    type: string
    x-oad-type: parameter
...
  Content:
    name: content
    in: body
    required: true
    schema:
      x-oad-type: string
      type: string
      title: content
    x-oad-type: parameter
  Timestamp:
    name: timestamp
    in: body
    required: true
    schema:
      x-oad-type: string
      type: string
      title: timestamp
      format: date
    x-oad-type: parameter
  Key:
    name: key
    in: body
    required: true
    schema:
      x-oad-type: string
      type: string
      title: key
    x-oad-type: parameter
...
  Tag:
    name: tag
    in: path
    required: true
    type: string
    x-oad-type: parameter
...
securityDefinitions:
  Authorization:
    description: User authentication token
    name: User Token
    in: header
    type: apiKey
info:
  title: Mark Services API
  version: 1.0.0
  description: Communicate with the Mark Services project
  contact:
    name: Jake Ferrante
    email: ferrantej@knights.ucf.edu
  license:
    name: MIT
    url: 'https://mit-license.org'
host: mark-services.azurewebsites.net
basePath: /api/v1
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
swagger: '2.0'